Refactor: change in climate structure
